:recycle: move interfaces/external to infrastructure/external
