+++ infrastructure/external/mock_external_e2e/mock_knoq.go
@@ -2,7 +2,7 @@ package mock_external_e2e //nolint:revive
 
 import (
 	"github.com/gofrs/uuid"
-	"github.com/traPtitech/traPortfolio/interfaces/external"
+	"github.com/traPtitech/traPortfolio/infrastructure/external"
 	"github.com/traPtitech/traPortfolio/usecases/repository"
 	"github.com/traPtitech/traPortfolio/util/mockdata"
 )
