+++ infrastructure/external/portal.go
@@ -1,4 +1,6 @@
-package infrastructure
+//go:generate go run github.com/golang/mock/mockgen@latest -source=$GOFILE -destination=mock_$GOPACKAGE/mock_$GOFILE
+
+package external
 
 import (
 	"encoding/json"
@@ -7,40 +9,49 @@ import (
 	"time"
 
 	"github.com/patrickmn/go-cache"
-	"github.com/traPtitech/traPortfolio/interfaces/external"
-	"github.com/traPtitech/traPortfolio/interfaces/external/mock_external_e2e"
 	"github.com/traPtitech/traPortfolio/util/config"
 )
+
+type PortalUserResponse struct {
+	TraQID         string `json:"id"`
+	RealName       string `json:"name"`
+	AlphabeticName string `json:"alphabeticName"`
+}
+
+type PortalAPI interface {
+	GetUsers() ([]*PortalUserResponse, error)
+	GetUserByTraqID(traQID string) (*PortalUserResponse, error)
+}
 
 const (
 	cacheKey = "portalUsers"
 )
 
-type PortalAPI struct {
+type portalAPI struct {
 	apiClient
 	cache *cache.Cache
 }
 
-func NewPortalAPI(conf *config.PortalConfig, isDevelopment bool) (external.PortalAPI, error) {
-	if isDevelopment {
-		return mock_external_e2e.NewMockPortalAPI(), nil
-	}
+func NewPortalAPI(conf *config.PortalConfig, isDevelopment bool) (PortalAPI, error) {
+	// if isDevelopment {
+	// 	return mock_external_e2e.NewMockPortalAPI(), nil
+	// }
 
 	jar, err := newCookieJar(conf.API(), "access_token")
 	if err != nil {
 		return nil, err
 	}
 
-	return &PortalAPI{
+	return &portalAPI{
 		apiClient: newAPIClient(jar, conf.API()),
 		cache:     cache.New(1*time.Hour, 2*time.Hour),
 	}, nil
 }
 
-func (a *PortalAPI) GetUsers() ([]*external.PortalUserResponse, error) {
+func (a *portalAPI) GetUsers() ([]*PortalUserResponse, error) {
 	portalUsers, found := a.cache.Get(cacheKey)
 	if found {
-		return portalUsers.([]*external.PortalUserResponse), nil
+		return portalUsers.([]*PortalUserResponse), nil
 	}
 
 	res, err := a.apiGet("/user")
@@ -52,7 +63,7 @@ func (a *PortalAPI) GetUsers() ([]*external.PortalUserResponse, error) {
 	if res.StatusCode != http.StatusOK {
 		return nil, fmt.Errorf("GET /user failed: %v", res.Status)
 	}
-	var userResponses []*external.PortalUserResponse
+	var userResponses []*PortalUserResponse
 	if err := json.NewDecoder(res.Body).Decode(&userResponses); err != nil {
 		return nil, fmt.Errorf("decode failed: %w", err)
 	}
@@ -60,7 +71,7 @@ func (a *PortalAPI) GetUsers() ([]*external.PortalUserResponse, error) {
 	return userResponses, nil
 }
 
-func (a *PortalAPI) GetUserByTraqID(traQID string) (*external.PortalUserResponse, error) {
+func (a *portalAPI) GetUserByTraqID(traQID string) (*PortalUserResponse, error) {
 	if traQID == "" {
 		return nil, fmt.Errorf("invalid traQID")
 	}
@@ -75,7 +86,7 @@ func (a *PortalAPI) GetUserByTraqID(traQID string) (*external.PortalUserResponse
 		return nil, fmt.Errorf("GET /user/%v failed: %v", traQID, res.Status)
 	}
 
-	var userResponse external.PortalUserResponse
+	var userResponse PortalUserResponse
 	if err := json.NewDecoder(res.Body).Decode(&userResponse); err != nil {
 		return nil, fmt.Errorf("decode failed: %w", err)
 	}
@@ -84,5 +95,5 @@ func (a *PortalAPI) GetUserByTraqID(traQID string) (*external.PortalUserResponse
 
 // Interface guards
 var (
-	_ external.PortalAPI = (*PortalAPI)(nil)
+	_ PortalAPI = (*portalAPI)(nil)
 )
