+++ infrastructure/external/traq.go
@@ -1,4 +1,6 @@
-package infrastructure
+//go:generate go run github.com/golang/mock/mockgen@latest -source=$GOFILE -destination=mock_$GOPACKAGE/mock_$GOFILE
+
+package external
 
 import (
 	"encoding/json"
@@ -6,29 +8,43 @@ import (
 	"net/http"
 
 	"github.com/gofrs/uuid"
-	"github.com/traPtitech/traPortfolio/interfaces/external"
-	"github.com/traPtitech/traPortfolio/interfaces/external/mock_external_e2e"
+	"github.com/traPtitech/traPortfolio/domain"
 	"github.com/traPtitech/traPortfolio/util/config"
 )
 
-type TraQAPI struct {
+type TraQUserResponse struct {
+	ID    uuid.UUID        `json:"id"`
+	State domain.TraQState `json:"state"`
+}
+
+type TraQGetAllArgs struct {
+	IncludeSuspended bool
+	Name             string
+}
+
+type TraQAPI interface {
+	GetUsers(args *TraQGetAllArgs) ([]*TraQUserResponse, error)
+	GetUser(userID uuid.UUID) (*TraQUserResponse, error)
+}
+
+type traQAPI struct {
 	apiClient
 }
 
-func NewTraQAPI(conf *config.TraqConfig, isDevelopment bool) (external.TraQAPI, error) {
-	if isDevelopment {
-		return mock_external_e2e.NewMockTraQAPI(), nil
-	}
+func NewTraQAPI(conf *config.TraqConfig, isDevelopment bool) (TraQAPI, error) {
+	// if isDevelopment {
+	// 	return mock_external_e2e.NewMockTraQAPI(), nil
+	// }
 
 	jar, err := newCookieJar(conf.API(), "r_session")
 	if err != nil {
 		return nil, err
 	}
 
-	return &TraQAPI{newAPIClient(jar, conf.API())}, nil
+	return &traQAPI{newAPIClient(jar, conf.API())}, nil
 }
 
-func (a *TraQAPI) GetUsers(args *external.TraQGetAllArgs) ([]*external.TraQUserResponse, error) {
+func (a *traQAPI) GetUsers(args *TraQGetAllArgs) ([]*TraQUserResponse, error) {
 	res, err := a.apiGet(fmt.Sprintf("/users?include-suspended=%t&name=%s", args.IncludeSuspended, args.Name))
 	if err != nil {
 		return nil, err
@@ -39,14 +55,14 @@ func (a *TraQAPI) GetUsers(args *external.TraQGetAllArgs) ([]*external.TraQUserR
 		return nil, fmt.Errorf("GET /users failed: %v", res.Status)
 	}
 
-	var usersResponse []*external.TraQUserResponse
+	var usersResponse []*TraQUserResponse
 	if err := json.NewDecoder(res.Body).Decode(&usersResponse); err != nil {
 		return nil, fmt.Errorf("decode failed: %w", err)
 	}
 	return usersResponse, nil
 }
 
-func (a *TraQAPI) GetUser(userID uuid.UUID) (*external.TraQUserResponse, error) {
+func (a *traQAPI) GetUser(userID uuid.UUID) (*TraQUserResponse, error) {
 	res, err := a.apiGet(fmt.Sprintf("/users/%v", userID))
 	if err != nil {
 		return nil, err
@@ -57,7 +73,7 @@ func (a *TraQAPI) GetUser(userID uuid.UUID) (*external.TraQUserResponse, error) 
 		return nil, fmt.Errorf("GET /users/%v failed: %v", userID, res.Status)
 	}
 
-	var userResponse external.TraQUserResponse
+	var userResponse TraQUserResponse
 	if err := json.NewDecoder(res.Body).Decode(&userResponse); err != nil {
 		return nil, fmt.Errorf("decode failed: %v", err)
 	}
@@ -66,5 +82,5 @@ func (a *TraQAPI) GetUser(userID uuid.UUID) (*external.TraQUserResponse, error) 
 
 // Interface guards
 var (
-	_ external.TraQAPI = (*TraQAPI)(nil)
+	_ TraQAPI = (*traQAPI)(nil)
 )
