+++ infrastructure/repository/user_impl_test.go
@@ -15,8 +15,8 @@ import (
 	"github.com/gofrs/uuid"
 	"github.com/stretchr/testify/assert"
 	"github.com/traPtitech/traPortfolio/domain"
-	"github.com/traPtitech/traPortfolio/interfaces/external"
-	"github.com/traPtitech/traPortfolio/interfaces/external/mock_external"
+	"github.com/traPtitech/traPortfolio/infrastructure/external"
+	"github.com/traPtitech/traPortfolio/infrastructure/external/mock_external"
 	"github.com/traPtitech/traPortfolio/usecases/repository"
 )
