+++ integration_tests/repository/contest_test.go
@@ -8,8 +8,8 @@ import (
 	"github.com/gofrs/uuid"
 	"github.com/stretchr/testify/assert"
 	"github.com/traPtitech/traPortfolio/domain"
+	"github.com/traPtitech/traPortfolio/infrastructure/external/mock_external_e2e"
 	"github.com/traPtitech/traPortfolio/integration_tests/testutils"
-	"github.com/traPtitech/traPortfolio/interfaces/external/mock_external_e2e"
 
 	irepository "github.com/traPtitech/traPortfolio/infrastructure/repository"
 	urepository "github.com/traPtitech/traPortfolio/usecases/repository"
