--- interfaces/external/traq.go
@@ -1,23 +1,0 @@
-//go:generate go run github.com/golang/mock/mockgen@latest -source=$GOFILE -destination=mock_$GOPACKAGE/mock_$GOFILE
-
-package external
-
-import (
-	"github.com/gofrs/uuid"
-	"github.com/traPtitech/traPortfolio/domain"
-)
-
-type TraQUserResponse struct {
-	ID    uuid.UUID        `json:"id"`
-	State domain.TraQState `json:"state"`
-}
-
-type TraQGetAllArgs struct {
-	IncludeSuspended bool
-	Name             string
-}
-
-type TraQAPI interface {
-	GetUsers(args *TraQGetAllArgs) ([]*TraQUserResponse, error)
-	GetUser(userID uuid.UUID) (*TraQUserResponse, error)
-}
